tests/wpt/web-platform-tests/css/css-ui/parsing/text-overflow-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().textOverflow</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#text-overflow">
<meta name="assert" content="text-overflow computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-overflow", "clip");
test_computed_value("text-overflow", "ellipsis");
</script>
</body>
</html>
